split benchmarks by implementation
